+++ dubbo-config/dubbo-config-api/src/test/java/org/apache/dubbo/config/dubboserver/DubboServerTest.java
@@ -63,9 +63,9 @@ public class DubboServerTest {
 
     @Test
     public void basicConfiguration() {
-        DubboServer dubboServer = DubboServer.getInstance();
-        assertEquals(dubboServer.isStarted(), false);
-        assertEquals(dubboServer.isInitialized(), false);
+//        DubboServer dubboServer = DubboServer.getInstance();
+//        assertEquals(dubboServer.isStarted(), false);
+//        assertEquals(dubboServer.isInitialized(), false);
 //        assertNotEquals(dubboServer.getApplication(), null);
 //        dubboServer.start();
